requirements.txt
requests
Flask
GitPython
aiohttp[speedups]
beautifulsoup4
cloudscraper
fastapi==0.104.1
gunicorn
mangum
uvicorn[standard]
flask-socketio
pytest
asyncclick
libtorrent
bs4
python-qbittorrent